+++ Weather App/SceneDelegate.swift
@@ -1,0 +1,19 @@
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo _: UISceneSession, options _: UIScene.ConnectionOptions) {
+        guard let _ = scene as? UIWindowScene else { return }
+    }
+
+    func sceneDidDisconnect(_: UIScene) {}
+
+    func sceneDidBecomeActive(_: UIScene) {}
+
+    func sceneWillResignActive(_: UIScene) {}
+
+    func sceneWillEnterForeground(_: UIScene) {}
+
+    func sceneDidEnterBackground(_: UIScene) {}
+}
