+++ Weather App/model/CitiesProvider.swift
@@ -5,19 +5,18 @@ class CitiesProvider {
         guard let path = Bundle.main.path(forResource: "city.list", ofType: "json") else {
             return []
         }
-        var cities: [City] = []
+        var decoded: [City] = []
         do {
             let data = try Data(contentsOf: URL(fileURLWithPath: path), options: .mappedIfSafe)
             let decoder = JSONDecoder()
-            let decoded = try decoder.decode(CitiesRepository.self, from: data)
-            cities = decoded.cities
-            cities = Array(Set(cities))
-            cities = cities.sorted { $0.name.lowercased() < $1.name.lowercased() }
+            decoded = try decoder.decode([City].self, from: data)
+            decoded = Array(Set(decoded))
+            decoded = decoded.sorted { $0.name.lowercased() < $1.name.lowercased() }
         } catch {
             // Error Handler
             print(error)
         }
 
-        return cities
+        return decoded
     }
 }
